added the basic logic of placing random tiles
gonna add the hand system and some very basic ui next commit
