+++ ThiefGame/Assets/Materials/TileTestColor.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 41497ff304f32d64187cf1cd1b9d296b
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
