+++ ThiefGame/Assets/Prefabs/Tiles/TestTile 2.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: d01d3b3c7c53c5a4eba0296ede492919
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
