+++ ThiefGame/Assets/Scenes/BasicTilePlacement Clone.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 95c3dc8d2c141824c9ff811498c9ac7f
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
