+++ ThiefGame/Assets/Scripts/Board.cs
@@ -16,6 +16,8 @@ public class Board : MonoBehaviour
     [HideInInspector]
     public static Dictionary<TileCoordinate, TileGhost> availableTiles = new();
 
+    private TileDeck deck;
+
     // Start is called before the first frame update
     void Start()
     {
@@ -24,6 +26,8 @@ public class Board : MonoBehaviour
             existingTiles.Add(t.coordinate, t);
         }
         ExpandFromTile(_initialTile.GetComponent<Tile>());
+
+        deck = GetComponent<TileDeck>();
     }
 
     // Update is called once per frame
@@ -33,7 +37,17 @@ public class Board : MonoBehaviour
     }
 
     public void AddTile(TileGhost tileGhostToReplace) {
-        GameObject newTile = Instantiate(this.newTile, new Vector3(tileGhostToReplace.coordinate.x, 0f, tileGhostToReplace.coordinate.y), Quaternion.identity);
+
+        //will get rid of the ghosting and what not later
+        //just a basic fix to stop errors 
+        if(deck.IsEmpty())
+        {
+            Debug.Log("deck empty");
+            return;
+        }
+        newTile = deck.DrawTile();
+        newTile = Instantiate(this.newTile, new Vector3(tileGhostToReplace.coordinate.x, 0f, tileGhostToReplace.coordinate.y), Quaternion.identity);
+
         newTile.name = "Tile (" + tileGhostToReplace.coordinate.x + "," + tileGhostToReplace.coordinate.y + ")";
         newTile.transform.SetParent(transform);
         Tile newTileComponent = newTile.GetComponent<Tile>();
