+++ ThiefGame/Assets/Scripts/Deck.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 916ea8914fd85f24fbe4b65b0c6b1b61
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
